API/CLI: split out the CLI into a new package (`nonos-cli`)
