+++ .github/workflows/build-and-publish.yml
@@ -1,0 +1,52 @@
+name: Build (and publish) package
+# this is a reusable workflow that takes a package name as an input,
+# builds it and *conditionally* publishes to PyPI
+
+on:
+  workflow_call:
+    inputs:
+      package-name:
+        type: string
+        required: true
+      publish:
+        type: boolean
+        required: false
+        default: false
+
+jobs:
+  build:
+    name: Build artifacts
+    runs-on: ubuntu-latest
+    steps:
+    - name: Checkout Source
+      uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
+    - uses: astral-sh/setup-uv@4959332f0f014c5280e7eac8b70c90cb574c9f9b # v6.6.0
+      with:
+        enable-cache: true
+        cache-suffix: build
+    - name: Build distributions
+      shell: bash -l {0}
+      run: uv build --package ${{ inputs.package-name }}
+    - uses: actions/upload-artifact@ea165f8d65b6e75b540449e92b4886f43607fa02 # v4.6.2
+      with:
+        name: artifacts-${{ inputs.package-name }}
+        path: dist
+
+  publish:
+    name: Download (and publish)
+    needs: [build]
+    runs-on: ubuntu-latest
+    environment:
+      name: pypi
+      url: https://pypi.org/p/${{ inputs.package-name }}
+    permissions:
+      id-token: write
+    steps:
+    - uses: actions/download-artifact@d3f86a106a0bac45b974a628896c90dbdf5c8093 # v4.3.0
+      with:
+        name: artifacts-${{ inputs.package-name }}
+        path: dist
+
+    - name: Publish package distributions to PyPI
+      if: inputs.publish
+      uses: pypa/gh-action-pypi-publish@76f52bc884231f62b9a034ebfe128415bbaabdfc # v1.12.4
