+++ .github/workflows/cd.yml
@@ -1,30 +1,45 @@
 name: CD
 on:
   push:
-    tags: v*
+    tags:
+    # use vMM.mm.pp to publish nonos (library) and cli-vMM.mm.pp to publish nonos-cli
+    - v*
+    - cli-v*
   pull_request:
     paths:
     - .github/workflows/cd.yml
+    - .github/workflows/build-and-publish.yml
+    - '*/**/pyproject.toml'
+    - CITATION.cff
+    - scripts/pre_publish_checks.py
 
 jobs:
-  pypi-publish:
-    name: Upload release to PyPI
+  pre-publish-checks:
     runs-on: ubuntu-latest
-    environment:
-      name: pypi
-      url: https://pypi.org/p/nonos
-    permissions:
-      id-token: write
+    name: pre-publication checks
     steps:
-    - name: Checkout Source
-      uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8 # v5.0.0
+    - uses: actions/checkout@08c6903cd8c0fde910a37f88322edcfb5dd907a8 # v5.0.0
+      with:
+        fetch-depth: 0
     - uses: astral-sh/setup-uv@4959332f0f014c5280e7eac8b70c90cb574c9f9b # v6.6.0
       with:
-        python-version: '3.13'
         enable-cache: false
-    - name: Build distributions
-      shell: bash -l {0}
-      run: uv build
-    - name: Publish package distributions to PyPI
-      if: github.repository == 'la-niche/nonos' && github.event_name == 'push' && startsWith(github.ref, 'refs/tags/v')
-      uses: pypa/gh-action-pypi-publish@76f52bc884231f62b9a034ebfe128415bbaabdfc # v1.12.4
+    - run: uv run scripts/pre_publish_checks.py
+
+  lib:
+    needs: [pre-publish-checks]
+    uses: ./.github/workflows/build-and-publish.yml
+    with:
+      package-name: nonos
+      publish: ${{ github.event_name == 'push' && startsWith(github.ref, 'refs/tags/v') }}
+    permissions:
+      id-token: write
+
+  cli:
+    needs: [pre-publish-checks]
+    uses: ./.github/workflows/build-and-publish.yml
+    with:
+      package-name: nonos-cli
+      publish: ${{ github.event_name == 'push' && startsWith(github.ref, 'refs/tags/cli-v') }}
+    permissions:
+      id-token: write
