+++ cli/pyproject.toml
@@ -1,0 +1,50 @@
+[build-system]
+requires = ["flit_core >=3.11,<4"]
+build-backend = "flit_core.buildapi"
+
+[project]
+name = "nonos-cli"
+version = "0.1.0"
+description = "A companion command line interface for nonos"
+readme = "README.md"
+authors = [
+    { name = "G. Wafflard-Fernandez" },
+    { name = "C.M.T. Robert" },
+]
+license = "GPL-3.0-only"
+license-files = ["LICENSE"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Environment :: Console",
+    "Framework :: Matplotlib",
+    "Intended Audience :: Science/Research",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Programming Language :: Python :: 3.14",
+    "Programming Language :: Python :: Implementation :: CPython",
+    "Topic :: Scientific/Engineering :: Visualization",
+]
+requires-python = ">=3.10"
+dependencies = [
+    "inifix>=5.1.0",
+    "loguru>=0.5.3",
+    "matplotlib>=3.5.0",
+    "nonos>=0.19.0",
+    "numpy>=1.21.2",
+    "termcolor>=2.1.0",
+    "tqdm>=4.64.1",
+]
+
+[project.scripts]
+nonos = "nonos_cli:main"
+
+[project.optional-dependencies]
+all = [
+    "nonos[all]",
+]
+
+[tool.uv.sources]
+nonos = { workspace = true }
