+++ cli/src/nonos_cli/__main__.py
@@ -1,6 +1,6 @@
 import sys
 
-from nonos.main import main
+from nonos_cli import main
 
 if __name__ == "__main__":
     sys.exit(main())
