+++ cli/src/nonos_cli/logo.txt
